streaming_response/requirements.txt
strands-agents
fastapi
uvicorn
openai==2.15.0
boto3==1.42.33
urllib3>=2.6.3,<3.0.0
xai_sdk==1.5.0